Updates after review, and split up FV tests
